Update PyuEye.py
Python bindings for uEye API (www.ids-imaging.com)

Description:
PyuEye is a lean wrapper implementation of Python function objects that represent uEye API functions. As there is no intelligence in the PyuEye interface, it is as near as possible to the uEye C API and always up-to-date. Benefit from many Python advantages, for example the interactive programming of your uEye camera. Write and test small code snippets without the complex setup of a programming IDE with a toolchain. The PyuEye interface is ideal for prototyping uEye camera applications for all supported uEye platforms (Windows, Linux, Linux Embedded).