Create Alg4_LaPlacianOperator.py
It computes the DFT of the image where a Laplacian filter (with sigma value 3) has been applied. The Laplacian filter is an edge detector used to compute the second derivatives of an image, measuring the rate at which the first derivatives change. This determines if a change in adjacent pixel values is from an edge or continuous progression.