Update Alg2_LaplacianOperator.py
Equivalent to Algorithm 2 but instead of having as an input the original image, it uses the image that has been modified by the Laplacian filter. A Laplacian filter is an edge detector used to compute the second derivatives of an image, measuring the rate at which the first derivatives change. This determines if a change in adjacent pixel values is from an edge or continuous progression.